Re-hide implementation and add a comment about the pimpl
